+++ recon/readable_sources/app/g1/set_delay_deadline.c
@@ -11,7 +11,7 @@
 /* Reconstructed FUN_0002a4d8 @ 0x2a4d8  (parity: 300/300 trials, PROVEN) */
 
 #include <stdint.h>
-extern uint64_t k_uptime_get_1(void);
+extern unsigned long long k_uptime_get_1(void);
 
 void set_delay_deadline(unsigned int param_1)
 {
