+++ recon/verified/src/FUN_00067600.c
@@ -4,7 +4,7 @@
 extern void FUN_0007e2ec(uint32_t, uint32_t);
 extern void FUN_0007e2fa(unsigned long, ...);
 extern int FUN_00085316(uintptr_t, void *);
-extern int FUN_0008539a(uintptr_t, uint32_t);
+extern unsigned int FUN_0008539a(unsigned long, unsigned int);
 
 struct callback_payload {
   uint8_t event;
